msm: ipahal: fields is non-null in ipahal_read_reg_n_fields()
All callers of ipahal_read_reg_fields() pass the address of an
object for the "fields" argument, so it will always be non-null.

Similarly, all callers of ipahal_read_reg_n_fields() (including
ipahal_read_reg_fields()) pass the (non-null) address of an object
as the "fields" argument.

So there is no need for ipahal_read_reg_n_fields() to verify fields
is non-null.

Change-Id: Idd3312c16456d34359ced43adcb74eb4b45b5437
Signed-off-by: Alex Elder <elder@codeaurora.org>