msm: ipa: modify _ipa_read_ep_reg_v3_0()
The debugfs file /sys/kernel/debug/ipa/ep_reg dumps the contents
of a set of registers for each endpoint.  One of the registers is
IPA_ENDP_INIT_ROUTE_n, which is no longer supported starting with
version 4.0 of IPA hardware.  We only support version 3.5.1 now, but
that will change.

The function that reads the registers is _ipa_read_ep_reg_v3_0().
Add a comment explaining that the read of that register should
be conditional when we support newer versions of IPA hardware.
And knowing this is coming, split up the scnprintf() call in that
function to make it easy to implement the conditional code.

Change-Id: Ib0ed54d745b025683e246119565b289eb76a868f
Signed-off-by: Alex Elder <elder@codeaurora.org>