msm: ipahal: fields is non-null in ipahal_write_reg_n_fields()
All callers of ipahal_write_reg_fields() pass the address of an
object for the "fields" argument, which will always be non-null.
This is a wrapper around ipahal_write_reg_n_fields(), and will
never pass it a null "fields" value.

Similarly, most callers of ipahal_write_reg_n_fields() also pass the
address of an object as the "fields" argument.  The only exceptions are:
  - ipa3_cfg_ep_status(), which passes "ep_status" as its "fields"
    value.  The only caller of this function passes the address of
    an object as its "ip_status" value.
  - ipa3_cfg_ep_metadata_mask() which passes "metadata_mask" as its
    "fields" value.  The only caller of this function passes the
    address of an object as its "metadata_mask" value.
  - ipa3_cfg_ep_ctrl() which passes "ep_ctrl" as its "fields" value.
    All callers of this function pass the address of an object as
    the "ep_ctrl" value.
  - ipa3_cfg_ep_aggr() which passes "ep_aggr" as its "fields" value.
    The only caller of this function passes the address of an object
    as its "ep_aggr" value.
  - ipa3_cfg_ep_holb() which passes "ep_holb" as its "fields" value.
    Both callers of this function pass the address of an object as
    the "ep_holb" value.
  - ipa3_write_rsrc_grp_type_reg() which passes "val" as its "fields"
    value.  Both callers of this function pass the address of an
    object as the "val" value.

So all code paths that lead to ipahal_write_reg_n_fields() pass the
address of an object as the "fields" argument, so it will never be
non-null.  So the function does not need to check this.

Change-Id: I0fdaa0ffb658a20c1d00a73a21654c0122461066
Signed-off-by: Alex Elder <elder@codeaurora.org>