msm: ipahal: don't put zero immcmd_obj on the stack
In ipahal_imm_cmd_init() there's an object defined whose purpose is
just to be zero-filled, for comparison against other objects.  Give
that object static lifetime rather than putting it on the stack.
This avoids consuming the stack space and has the side-effect of
being initially zero-filled.  Make it const as well.

Change-Id: I492de3717886726223f087a9b01e17b6d987481e
Signed-off-by: Alex Elder <elder@codeaurora.org>