msm: ipahal: populate a single imm_cmds structure
If a system has an IPA, it has exactly one of them, and it will have
a single hardware version/type.  There is no need to populate the
entire array of immediate command objects for all hardware versions.
Instead, we can use a single block of immediate commands, which can
be built up using the ipahal_imm_cmd_objs[][] array.  We still
verify every entry in this immediate command block, but only those
used are checked.

For now we just populate this array; this commit temporarily
includes a test to verify the result is the same as what the
existing code produces for the current version as well.

Change-Id: Icd94dbdf2db97a46f26da09f4df33f7d0b39e004
Signed-off-by: Alex Elder <elder@codeaurora.org>