msm: ipahal: define and use OPCODE_INVAL
Currently an opcode value of -1 in an ipahal_imm_cmd_objs[][] entry
represents a immediate command that is not valid for a particular
version of IPA hardware.  This patch defines symbolic constant
OPCODE_INVAL for that purpose.  The value used is the same, but it's
defined as an explicitly unsigned all-1's value rather than having -1
be implicitly converted to "opcode" field's unsigned type.

Change-Id: Id71246c3676abf0d5e460e7fd8dc9b1e7efea29e
Signed-off-by: Alex Elder <elder@codeaurora.org>