msm: ipahal: kill ipahal_imm_cmd_validate()
There's little value in ipahal_imm_cmd_validate(), which simply
checks for a null construct pointer in an immediate command object.
We can see by inspection that no valid entry in ipahal_imm_cmd_objs[][]
has a null "construct" value, so there's no need to check this at
runtime.  So get rid of the call, and the function.  Add a BUG_ON()
in place of the validation call; if it ever occurs it will print
plenty of information to diagnose the problem.

Change-Id: I40c3896819b0a5cb2c18c8611fef81f4d3def492
Signed-off-by: Alex Elder <elder@codeaurora.org>