msm: ipahal: reformat ipahal_imm_cmd_objs[][]
Reformat the ipahal_imm_cmd_objs[][] array to more obviously show
which groups of definitions constitute sub-arrays.  Each major index
value in the array has a block of initializers for the entries in that
index's sub-array within ipahal_imm_cmd_objs[][].

Beyond that, lay out the initializers in a way that's typical in
upstream code.

Change-Id: I8c2a3fce9d321ff34acfefe454a26dd7eda44550
Signed-off-by: Alex Elder <elder@codeaurora.org>