msm: ipahal: return 0 for errors in ipahal_imm_cmd_get_opcode*()
The value returned by ipahal_imm_cmd_get_opcode() is unsigned.
However it does some error checking, and if an error occurs it
returns negative errno.

None of the callers of this function actually check for this.
(And we can most likely just do away with all of these checks
because the conditions they're checking will never occur.)

However, for now, let's at least return a non-negative value.  Have
this function return 0 on error.  It's no worse, and probably better
than the -EFAULT (-14) as it does right now.

Change-Id: I1a465404bd7b1d32aa4162790f8e8a25be96cd9c
Signed-off-by: Alex Elder <elder@codeaurora.org>