msm: ipahal: cmd is valid in ipahal_construct_imm_cmd()
The 20 callers of ipahal_construct_imm_cmd() "ipahal.c" all provide
a "cmd" value that is a valid member of the ipahal_imm_cmd_name
enumerated type.  So there is no need for ipahal_construct_imm_cmd()
to check for this condition.  Get rid of that check, and rely on
developers to always provide valid "cmd" values.

Change-Id: I469702dc6f500ba59988f58551eb1c4beac27886
Signed-off-by: Alex Elder <elder@codeaurora.org>