msm: ipahal: skip_pipeline_clear is false in ipahal_construct_nop_imm_cmd()
The first argument to ipahal_construct_nop_imm_cmd() is a flag
indicating what value to supply for the skip_pipeline_clear option
in the register_write immediate command.  All callers supply false
for this, so we can just get rid of the argument and use false in
its place.

Change-Id: I5b72d3b2ebc1d16c8271517067ebc254d3d869ca
Signed-off-by: Alex Elder <elder@codeaurora.org>