msm: ipahal: assume clear option in ipahal_construct_nop_imm_cmd()
There is only one caller of ipahal_construct_nop_imm_cmd(), and it
supplies IPAHAL_FULL_PIPELINE_CLEAR as the pipline_clr_opt value.
Just assume this, and drop the pipline_clr_opt parameter entirely.

When supporting future functionality we will need to add this
parameter back to the function, but it's better to keep things as
simple as possible for now.

Change-Id: Id4809455718b5898d5ddda5660a291b90a8f79b9
Signed-off-by: Alex Elder <elder@codeaurora.org>