msm: ipahal: kill the ipahal_pkt_status_objs[] table
The ipahal_pkt_status_objs[] table is designed to function the same
way the register object and immediate command object tables do,
allowing different versions of IPA hardware to parse status packets
differently.  The fact is, there is no variability across hardware
versions in how this is done, so at least for now we don't need this
extra layer of abstraction here.

So get rid of the ipahal_pkt_status_objs[], as well as the type of
its elements.  Delete the code that performed initialization of the
table, and update the description of ipahal_pkt_status_init().

Change-Id: I225a3922775132da289108676e7b27ae0c623c8c
Signed-off-by: Alex Elder <elder@codeaurora.org>