msm: ipahal: opcode is valid in ipahal_imm_cmd_get_opcode()
There is only one opcode that has an opcode of OPCODE_INVAL in the
ipahal_imm_cmds[] array.  That entry, for IPA_IMM_CMD_NAT_DMA, is
never used in the IPA lite code.  Since we know the opcode found in
the array by ipahal_imm_cmd_get_opcode() is never OPCODE_INVAL, we
can eliminate the check for that.

Change-Id: I99c18247f1b8b9f9aa1fb81523d33632f8bae8d7
Signed-off-by: Alex Elder <elder@codeaurora.org>