msm: ipahal: no need to cast params
There's no need to cast pointers to or from (void *).  The "params"
argument in every immediate command construct function is a void
pointer, so we can eliminate the casts in the assignments of typed
local variables to the value of the "params" argument.

Note that the previous casts were also eliminating the "pointer to
const" attribute of the "params" argument.  We now preserve that by
adding it to the definition of each of the affected local variables.

Change-Id: Ibac6620575cdabcefaa191d95e080e01a2ee9203
Signed-off-by: Alex Elder <elder@codeaurora.org>