msm: ipahal_fltrt: don't put zero fltrt_obj on the stack
In ipahal_fltrt_init() there's an object defined whose purpose is
just to be zero-filled, for comparison against other objects.  Give
that object static lifetime rather than putting it on the stack.
This avoids consuming the stack space and has the side-effect of
being initially zero-filled.  Make it const as well.

Change-Id: I97a57ff2e837d5c0accb0ac0c09b84bcc82ebc6f
Signed-off-by: Alex Elder <elder@codeaurora.org>