msm: ipahal_fltrt: populate a single fltrt structure
If a system has an IPA, it has exactly one of them, and it will have
a single hardware version/type.  There is no need to populate the
entire array of filter table objects for all hardware versions.
Instead, we can use a single block filter table object, which can be
built up using the ipahal_fltrt_objs[] array.  We still verify
this filter table, but only the used is checked.

For now we just populate the single table.  This commit temporarily
includes a test to verify the result is the same as what the
existing code produces for the current hardware version as well.

Change-Id: I9eab5e00594531e48c8937e8aca7791f721728df
Signed-off-by: Alex Elder <elder@codeaurora.org>