msm: ipahal_fltrt: use new ipahal_fltrt structure
All references to the ipahal_fltrt_objs[] array use the hardware
type found in the IPA HAL context as the index.

The ipa_fltrt was built up to contain exactly the same thing as that
sub-structure (i.e., the portion with that index value).

Replace all references to ipahal_fltrt_objs[ipahal_ctx->hw_type]
with references to ipahal_fltrt instead.

Change-Id: Id467a17877fd63b14275dbea4f6436e36f95911b
Signed-off-by: Alex Elder <elder@codeaurora.org>