msm: ipahal_fltrt: make ipahal_fltrt_objs[] constant
We now use the ipa_fltrt array throughout the HAL code.  The
ipahal_fltrt_objs[] array is used only for initializing that, and
this is done prior to populating zero entries found therein.  We no
longer need to update zero entries, so stop doing that.

That means ipahal_fltrt_objs[] is no longer ever modified, so we can
declare its content constant.

Change-Id: Ib7e1f55abfd5ba5f68f5160ba93257fc0c5fc96b
Signed-off-by: Alex Elder <elder@codeaurora.org>