msm: ipahal_fltrt: sysaddr_alignment is non-zero
We can see by inspection that all valid filter/route table entries
contain non-zero values for their sysaddr_alignment fields.  Don't
bother validating this condition at runtime; this is a constraint
for table entries and it will be the responsibility of the developer
to ensure it is satisfied.  (Some comments will be added later to
make this and other requirements explicit.)

Change-Id: Iab1fe8cc17bf608c9c61f7ec9ae0a1199ddfee08
Signed-off-by: Alex Elder <elder@codeaurora.org>