msm: ipa: move mmio setup into main driver
The only things required to set up ipa3_ctx->mmio are for the
ipa_wrapper_base, ipa_wrapper_size, and ctrl->ipa_reg_base_ofst
fields in the IPA context to be initialized.  All of this is done
relatively early in ipa3_plat_drv_probe() when probing for the main
IPA platform driver.  So move the ioremap() call for ipa3_ctx->mmio
into that probe routine, as soon as we have what we need.

Change-Id: I306c41a1debda0316d387dd7da83d7444c9517e2
Signed-off-by: Alex Elder <elder@codeaurora.org>