msm: ipahal: return NULL for bad params
If a construct command gets bad input parameters, the constructed
command is an error.  The caller is not adhering to the interface
the construct function honors, so the constructed command cannot
correctly implement what was requested.  Currently the code always
warns when this occurs, but sometimes the function proceeds anyway.

Callers of construct functions already must be prepared to handle a
memory allocation error, so we can always return a null pointer if
another error occurs.   So change all construct functions so that in
addition to warning about bad parameters, they return NULL rather
than a valid pointer.

Change-Id: Id4823d1d1aecfd0965e5e29fc23d4aaff788035e
Signed-off-by: Alex Elder <elder@codeaurora.org>