msm: ipahal: no cast on hw_status pointer
In ipa_pkt_status_parse() the hw_status pointer is assigned with a
cast, and in the process we lose the const attribute that comes from
the passed-in unparsed_status pointer.  There is no need to cast
assignments to void pointer variables, so make hw_status a pointer
to const and drop the cast in the assignment.

Change-Id: Ifb06b7a5cfe8a400f2d05df362a50c876b325653
Signed-off-by: Alex Elder <elder@codeaurora.org>