msm: ipahal: encapsulate exception type mapping
Create and use exception_map() to convert an exception type supplied
by the hardware into the a value used to indicate that type using
the HAL API.  IPAHAL_PKT_STATUS_EXCEPTION_MAX is used to indicate
an unrecognized exception type.

NOTE:  The effect on error is that EXCEPTION_NONE is indicated to
the caller of ipahal_pkt_status_parse().  This may not be desirable,
but it's the way the code behaved prior to this change.

Change-Id: I025d180612beb7b9f2cd5633441286c9890994b4
Signed-off-by: Alex Elder <elder@codeaurora.org>