msm: ipahal: use mask values in ipahal_pkt_status_mask
Rather than defining the ipahal_pkt_status_mask enumerated type to
have shift values (i.e., the amount by which a 1 should be shifted
left to get the desired value), just define the members of the type
to be mask, and assign their numeric values accordingly.  Drop the
"_SHFT" suffix from each name to reflect the new role.

This affects the definitions of IPAHAL_PKT_STATUS_MASK_FLAG_VAL()
and IPA_PKT_STATUS_SET_MSK(), but the changes are straightforward.

Change-Id: Ia4356bfac09214bf6039ca006a3c3c591d76cccd
Signed-off-by: Alex Elder <elder@codeaurora.org>