msm: ipahal: kill IPA_PKT_STATUS_SET_MSK()
IPA_PKT_STATUS_SET_MSK() is not a well-defined macro, assuming the
existence of variables "status" and "hw_status" in the environment
in which they're called.  But aside from that, its purpose is simply
to set a bit in a one variable (status) if a specified bit is set in
another (hw_status).  That simple activity is easily understood but
is obscured by the use of the macro.

Get rid of the macro, and open-code what it does where it's used.

Change-Id: I8b765f41257201369e70339eddac96181533dadc
Signed-off-by: Alex Elder <elder@codeaurora.org>