msm: ipahal: assign status mask directly in ipa_pkt_status_parse()
In ipa_pkt_status_parse() there is a large block of if statments
that set bits in the status field of an ipahal_pkt_status structure
based on bits set in status value returned from hardware.

The hardware bit values checked are exactly the same as the "soft"
status_mask values that will be returned, and the entire set of 16
possible status bits are processed.  So this entire block can be
simplified by just directly assigning the status_mask to the value
returned by the hardware.  The hardware field is 16 bits wide, so
and the status structure being filled will have already been zeroed
by the caller (ipa_pkt_status_parse()), so we don't need to mask the
result either.

If future versions of hardware change the format of the status bits
returned, we may have to do some extra mapping work, but for now
there is no need, and this is much simpler.

Change-Id: Iad07b309e69993b0d2c0024426951d94e1a61f23
Signed-off-by: Alex Elder <elder@codeaurora.org>