msm: ipahal: explicitly mask offset bits
In ipa_imm_cmd_construct_register_write_v_4_0() the offset field of
the payload (which is 16 bits wide) is assigned the input offset
value (which is 32 bits wide).  The language ensures the right thing
happens, but it's comforting to see the upper bits explicitly masked
off in such a case, to make it obvious this is occurring.

Currently, this has no effect, because a check earlier in the
function already rejects anything with an offset that can't be
represented in 16 bits.  The hardware supports 20 bits though, and
at some point that might be allowed...

Change-Id: I6e4dba5be0cbbf1bfba74c048b269b12be50bcf1
Signed-off-by: Alex Elder <elder@codeaurora.org>