msm: ipahal: clever macro magic
The command identifiers and function names used to define the
ipahal_imm_cmd_objs[][] array follow strict naming patterns.
We can take advantage of this and and define some macros that make
the initialization of the table much cleaner, by having the C
preprocessor fill in the common prefix portions of names.

Change-Id: I7c0631cc10ff9d0a12cc66424d3771990244ba13
Signed-off-by: Alex Elder <elder@codeaurora.org>