msm: ipahal: embed immediate command name in main table
Rather than having a separate table mapping immediate
command identifiers to their names, just add a name field
to the ipahal_imm_cmd_obj structure, and initialize it when
constructing the main ipahal_imm_cmd_objs[][] table.

The only user of the string table is ipahal_construct_imm_cmd(),
so update that to reflect the change.

Change-Id: I1aab6a2948fb1291b09a2081d06abc574416271a
Signed-off-by: Alex Elder <elder@codeaurora.org>