msm: ipahal_reg: make n_ofst a 16-bit value
All the "n_ofst" multipliers in the ipahal_reg_objs[][] table are
less than 65536, so can fit in a 16 bit field, so the type of the
"n_ofst" field to u16.  The only place it's used its value will
automatically be converted to u32.

Update all the initializers for that field to use 0x%04x format.

Change-Id: I1a9674f8c18e29802f1755ec7925be80f9c1d12f
Signed-off-by: Alex Elder <elder@codeaurora.org>