msm: ipahal_reg: don't cast construct fields pointer
In the IPA HAL register construct functions there is a "fields"
parameter that has void pointer type.  Within functions that need
it, a local variable of the appropriate type is assigned that
"fields" pointer value.  Void pointers do not need a cast when they
are assigned to another pointer type, so get rid of all those casts.

Beyond being unnecessary, the casts were removing the const qualifier
that the parameter had.  We can (must) change the local variables to
have that qualifier as well, to preserve the guarantee of the API.

Change-Id: I564c8d287b94d0f96964db40d8dbe5313189dafb
Signed-off-by: Alex Elder <elder@codeaurora.org>