msm: ipahal_reg: don't cast parse fields parameter
In the IPA HAL register parse functions there is a "fields"
parameter that has void pointer type.  Within functions that need
it, a local variable of the appropriate type is assigned that
"fields" pointer value.  Void pointers do not need a cast when they
are assigned to another pointer type, so get rid of all those casts.

Change-Id: Iebcfc09ebfd1e73eabf10560cacaeded70b68514
Signed-off-by: Alex Elder <elder@codeaurora.org>