msm: ipahal_reg: have construct functions return value
Register construct functions format a 32-bit register value.

Currently, construct functions take the address of a 32-bit object
and have a void return type.  It's not obvious that the pointed-to
object is initially 0, and the interface leaves open the possibility
that the pointer provided could be invalid.  (We can see by
inspection this is not the case, of course.)

Rather than passing the address of a 32-bit variable to receive the
result of a construct function, have the function just return the
result, building it up using a zero-initialized local variable.

Change-Id: If840643aa239434a40af6830e2cf11f9ee9e2afe
Signed-off-by: Alex Elder <elder@codeaurora.org>