msm: ipahal_reg: shorten IPA_ROUTE register field defns
The symbol names specifying the position and mask for register
fields defined in "ipahal_reg_i.h" are unwieldy.  They are so long
they make it difficult to write concise code.

This patch is a first in a series that begins to shorten these by
removing the register name that each field is associated with.
Almost all of the shortened symbol names are still unique, and
those that are not can be fairly obviously modified to disambiguate
them.

This patch shortens symbols names related to the IPA_ROUTE register.

Change-Id: Idbc76bc58c91cf4197c4a48b27321dbb11771f33
Signed-off-by: Alex Elder <elder@codeaurora.org>