msm: ipahal_reg: IPA_DEBUG_CNT_CTRL_n duplication
The symbols that specify the fields in the IPA_DEBUG_CNT_CTRL_n
register incorporate "DBG_CNT_" in their names, which is somewhat
redundant.  Shorten most of these by deleting that portion of their
name.  Leave it in the "TYPE" and "EN" fields, to ensure what
remains is still meaningful.

Change-Id: If9264d165e7ebff95ab67b75d39cf346ff9a9206
Signed-off-by: Alex Elder <elder@codeaurora.org>