msm: ipahal_reg: introduce field_mask abstraction
Explain how the bitmasks used in defining register fields, in the
form they're defined in "ipahal_reg_i.h", supply enough information
to determine both the width of a field and its position, making the
definition of a field's shift value unnecessary.

Define such a mask as a "field_mask" and create trivial macros to
compute the width and shift for a field using compiler built-in
functions.

Introduce field_gen(), which shifts a given value so into a field's
position so it can be recorded (using a bitwise OR operation) into
a register.  This function warns if the value exceeds what can be
represented in the field.  This does the same thing as IPA_SETFIELD(),
so redefine that macro using field_gen().

Introduce field_val(), which extracts the value in a field from a
register value provided.  Redefine the IPA_GETFIELD_FROM_REG() macro
using this new function.

Change-Id: Id0a51b144001f94eaea7786179fc1a3624c16fae
Signed-off-by: Alex Elder <elder@codeaurora.org>