msm: ipa: fix some initialization values
Change 6 values that exceed the width of the fields in registers
they are written to.

The ipa3_rsrc_src_grp_config[][] array is used to initialize some
register values in ipa3_set_resorce_groups_min_max_limits().  Each
entry defines a minimum and maximum value to use, for different
resource group types and values within those groups, for sources and
destinations.  These values are written into fixed-width fields.

There are 6 maximum values present in the array that are bigger than
what can be represented in the fields they are written into.  Reduce
those values so they are the maximum representable value instead.

Change-Id: Ife1eb2dfc49c7eb255c25fc613faa6378e35fb6b
Signed-off-by: Alex Elder <elder@codeaurora.org>