msm: ipahal_reg: fix bug in ipahal_get_disable_aggr_valmask()
The values returned in the "val" and "mask" fields of the structure
whose address is passed to ipahal_get_disable_aggr_valmask() are
computed "directly" rather than using IPA_GETFIELD_FROM_REG().
The code makes an incorrect assumption about the position of set
bits in AGGR_FORCE_CLOSE_BMSK, and as a result, the "val" returned
is wrong.

Fix the bug by having the "val" and "mask" values computed using
IPA_GETFIELD_FROM_REG(), passing it all-0 or all-1 register values
to set or clear the returned fields.

Change-Id: I1bce0152a5445b2c0f3cf1c2581b46dbb0d27cf0
Signed-off-by: Alex Elder <elder@codeaurora.org>