msm: ipahal_reg: fix bug in ipahal_get_status_ep_valmask()
The code in ipahal_get_status_ep_valmask() assumes the BMSK value is
not pre-shifted, and as a result it is masking off the wrong bits in
the value.  It is furthermore shifting the BMSK value again by the
SHFT value, which is also wrong.  Fix both of these problems.

Change-Id: Ia011bb916caa0f24268185fbc3474618afb1f432
Signed-off-by: Alex Elder <elder@codeaurora.org>