msm: ipahal_reg: valmask is good in ipahal_get_status_ep_valmask()
The only caller of ipahal_get_status_ep_valmask() passes the address
of an object as its valmask parameter, so it will never be null.
There is no need to check for a null pointer in this function; it is
up to the caller to ensure a valid pointer is provided.

Change-Id: I083d5ae3d23f8ffb2ed1c8544f5bc85c19ace6e2
Signed-off-by: Alex Elder <elder@codeaurora.org>