msm: ipahal_reg: fix QCNCM immediate command bugs
The construct and parse functions for the QCNCM immediate commands
handle the undefined fields in the register incorrectly.  It
probably doesn't matter inpractice, because they affected bits are
unused, but this patch fixes the problem anyway.

Previously the parse function would extract both sets of undefined
bits and save them to one "undefined" field in the ipahal_reg_qcncm
structure.  But the corresponding code in the construct function
would apply MODE_VAL_BMSK to what was in the "undefined" field,
zeroring all those set bits and effectively ignoring the register.
(It also would overwrite the mode_val field if a bad value was
supplied for "undefined"; fortunately the only caller basically does
a read/modify/write of the register.)

Replace the single "undefined" field in the ipahal_reg_qcncm
structure with two--one for each of its two "undefined" fields.
Use them to extract those values in the parse function, and write
them back in the construct function.

Change-Id: I79f7268dad6c64a9362730aded0ab385c6b1d463
Signed-off-by: Alex Elder <elder@codeaurora.org>