msm: ipahal_reg: drop ipa_assert() call in ipareg_construct_tx_cfg()
In ipareg_construct_tx_cfg() there is a check to ensure the
two tx*_prefetch_almost_empty_size fields are equal, and if they're
not, ipa_assert() is called.  This is the construct function for the
IPA_TX_CFG register.

Nothing references that register in the IPA lite code.  In the full
code there are two users, and both perform a read/modify/write of
the register.  The parse function already assigns the tx1 value
to have be the same as what was extracted for tx0, so we know the
two will have the same value.

There is therefore no need to verify this condition, unless the code
changes and a new caller changes these values so they differ.
Ultimately the tx1 value is ignored in the construct function, so it
doesn't affect what gets sent to the hardware.

Change-Id: I06e15d85c23e7bb79dae5a6edd9fefa653ddbdf9
Signed-off-by: Alex Elder <elder@codeaurora.org>