msm: ipahal_reg: offset cannot be 0
Ghanim Fodi, while reviewing this commit:
  msm: ipahal: 0 offset means uninitialized entry
pointed out that iparegs[] will never contain an entry with a zero
offset value.  So there's no need to check for that condition after
getting the offset as was done in that original patch.

The current code has boiled all of that down to a single place that
looks up the offset value, in ipahal_reg_n_offset().  Change the
BUG_ON() call there to no longer check for a zero offset value.

Change-Id: I1db83ef5383c42e00f2d06a3d21b20834e4d71e7
Signed-off-by: Alex Elder <elder@codeaurora.org>