msm: ipahal: tbls_num is non-zero in ipahal_rt_generate_empty_img()
There is no need to check whether the "tbls_num" value passed to
ipahal_rt_generate_empty_img() is non-zero, because we can check
that at runtime.

There are three callers of ipahal_rt_generate_empty_img():
  - _ipa_init_rt4_v3() passes ipa3_mem(V4_RT_NUM_INDEX) as "tbls_num"
  - _ipa_init_rt6_v3() passes ipa3_mem(V6_RT_NUM_INDEX)
  - ipa3_q6_clean_q6_rt_tbls() passes (hi - lo + 1), where "hi" and
    "lo" are either:
      - ipa3_mem(V4_MODEM_RT_INDEX_LO) and ipa3_mem(V4_MODEM_RT_INDEX_HI)
      - ipa3_mem(V6_MODEM_RT_INDEX_LO) and ipa3_mem(V6_MODEM_RT_INDEX_HI)

The previous patch added compile-time checks that ensure all of the
above values are non-zero.  For the third case, we know (lo <= hi)
and so (hi - lo + 1) is 1 or more.

Change-Id: I10aec756fd7241b01579581bf9f7f0daf15b8445
Signed-off-by: Alex Elder <elder@codeaurora.org>