msm: ipahal: fix bugs in ipa3_q6_clean_q6_flt_tbls()
In ipa3_q6_clean_q6_flt_tbls(), the value to pass for the header
sizes to ipahal_flt_generate_empty_img() is determined based on
whether HASHABLE or NON_HASHABLE is indicated in the parameters.
Despite this, the values passed are wrong--they're the same for both
in all cases, while there should be one indicating the HASHABLE
header size and a second the NON_HASHABLE one.  Fix this.

Change-Id: I5199fb865e801b9a9b69e03ad90ff0263f444a96
Signed-off-by: Alex Elder <elder@codeaurora.org>