msm: ipahal: remove a check in ipahal_rt_generate_empty_img()
The previous patch ensured certain conditions are met as soon as the
information needed to check them is available.  In particular, it
verified that the space set aside for non-hashed route tables is
sufficient to hold the number of entries required.  The specific
checks made allow this patch to remove a corresponding check in
ipahal_rt_generate_empty_img().

Three sites call ipahal_rt_generate_empty_img():
  - _ipa_init_rt4_v3() passes ipa3_mem(V4_RT_NUM_INDEX) as the value
    for tbls_num and ipa3_mem(V4_RT_NHASH_SIZE) as nhash_hdr_size.
  - _ipa_init_rt6_v3() passes ipa3_mem(V6_RT_NUM_INDEX) as the value
    for tbls_num and ipa3_mem(V6_RT_NHASH_SIZE) as nhash_hdr_size.
  - ipa3_q6_clean_q6_rt_tbls() passes values that depend on whether
    IPv4 or IPv6 entries are being cleaned.  But tbls_num and
    nhash_hdr_size are based on the values in the mem_partition[]
    array.
The previous patch ensures that, for the conditions presented by
all three of these callers, the nhash_hdr_size will be greater than
the size required for tbls_num route table entries.  So we no longer
need to check this in ipahal_rt_generate_empty_img().

Change-Id: I976ffa51c6126fe25ce4b7ff81d2f60a0605aa1c
Signed-off-by: Alex Elder <elder@codeaurora.org>