msm: ipahal: add early checks for hashed route table size
Verify in config_valid() that the space allotted for the hashed
route tables are big enough to hold the number of entries specified
for IPv4 and IPv6, including the slots reserved for the modem.

These checks match the ones done by ipahal_rt_generate_empty_img(),
using the values that the three callers of that function provide.
With these things verified once, early, we no longer need to check
them in every call to that function (addressed in the next patch).

Change-Id: Ic64fdb195af420f7480bff5677ef563e089f8326
Signed-off-by: Alex Elder <elder@codeaurora.org>