msm: ipahal: tbls_num is non-zero in ipahal_flt_generate_empty_img()
In ipahal_flt_generate_empty_img() there is a check to ensure the
tbls_num parameter is non-zero.  The previous patch ensures that
this is the case for two of the three of that function's callers,
and the third passes value 1.  So there's no need to check for that
condition in ipahal_flt_generate_empty_img().

Change-Id: I2f618de3d8bec967821c027e8d954e3d1dc51937
Signed-off-by: Alex Elder <elder@codeaurora.org>