msm: ipahal: add early checks for non-zero filter table size
Verify in config_valid() that the size of both the hashed and
unhashed filter tables are non-zero.

These checks match some done by ipahal_flt_generate_empty_img(),
using the values that the three callers of that function provide.
With these things verified once, early, we no longer need to check
them in every call to that function (addressed in the next patch).

Change-Id: I0f6f9e30c1c3ebb6ce5e5f2748233649b7253603
Signed-off-by: Alex Elder <elder@codeaurora.org>