msm: ipahal: header sizes are non-zero in ipahal_flt_generate_empty_img()
In ipahal_flt_generate_empty_img() there are checks to ensure the
hash_hdr_size and hash_hdr_size parameters are non-zero.  The
previous patch ensures that this is the case for all three of that
function's callers, so there's no need to check for that
condition in ipahal_flt_generate_empty_img().

Change-Id: Ic530fb2c8e8ab4ca6dd54923da6bf650d4ccc389
Signed-off-by: Alex Elder <elder@codeaurora.org>