msm: ipahal: update ipahal_flt_generate_empty_img()
There is a check in ipahal_flt_generate_empty_img() to ensure
there's sufficient space for the number of table entries begin
generated.  The size is compared in a way that differs from a
very similar check in ipahal_rt_generate_empty_img().  Change
the filter table check so it follows the pattern used for the
route table.

Change-Id: I072e7ce7e3f7c1917a2d783a868fd40f3fcec213
Signed-off-by: Alex Elder <elder@codeaurora.org>