msm: ipa: use a preallocated IPA HAL context
Use a statically-defined structure to hold the IPA HAL context.
Don't update all the references to ipahal_ctx; just have that
pointer refer to the static structure.  This eliminates a possible
(though unlikely) error in ipahal_init(), and makes it unnecessary
to free or reinitialize the pointer in error paths.

Change-Id: I62998d9cede792f3f6a38dd939a5ebeff3498d03
Signed-off-by: Alex Elder <elder@codeaurora.org>