msm: ipahal: cache empty table address
Every time the allocated empty filter/route table is used, its
address in system space is used, and that's always computed the same
way.  Compute that once, at allocation time.

This trades a small amount of space for a small amount of time, but
the purpose of the change is more about code readability.

Change-Id: I47eef1eb6f74c7dfbc3347804d8398840444069b
Signed-off-by: Alex Elder <elder@codeaurora.org>