msm: ipahal: introduce ipahal_dma_alloc()
Define a new interface ipahal_dma_alloc() for allocating DMA memory.
(It is basically a renamed ipahal_alloc_empty_img().)  The result of
the allocation is placed in the ipa_mem_buffer structure whose
address is provided, and the because the zalloc variant is used, the
allocated memory will be zeroed Returns 0 if successful, -ENOMEM
otherwise.

Also define ipahal_dma_free() to releaser previously-allocated DMA
memory.

Change-Id: I90b50a954eba722c831e83143cfa78ad4f726abd
Signed-off-by: Alex Elder <elder@codeaurora.org>