msm: ipa: use IPA HAL DMA functions in ipa3_init_smem_region()
Use ipahal_dma_alloc() and ipahal_dma_free() rather than the "raw"
DMA functions in ipa3_init_smem_region().  There is no need for the
caller to zero the allocated space (nor the mem structure it passes
in).  (Also: move a few memset() calls of structures closer to where
the structures are first used.)

The primary thing to address here is ensuring the device structure
used in the allocation is the same after the change as before.  The
structure used by ipahal_dma_alloc() is ipahal_ctx->ipa_pdev.  That
is set by ipahal_dev_init() to be the device structure for the AP
SMMU device.  A copy of that device pointer is in the context
block structure for that SMMU, in ipa3_ctx->ap_smmu_cb.dev.  And
that is exactly the pointer ipa3_init_smem_region() previously
provided for the DMA allocation request.

Change-Id: Ie2defd116df25378badaf2f1f1b7850105ff9380
Signed-off-by: Alex Elder <elder@codeaurora.org>