msm: ipahal: only system filter/route table entry addresses
The create_tbl_addr() function associated with filter/route tables
has a Boolean "is_sys" parameter to indicate whether the address is
to be converted to a system address, or to a local address.  The
current code only ever passes true for this parameter, and Ady says
support for local memory tables can be removed.  So get rid of the
"is_sys" parameter and simplify the function accordingly.

Change-Id: I8d12cab9130101bd0d8090731ee902f22da2f35a
Signed-off-by: Alex Elder <elder@codeaurora.org>