msm: ipa: introduce client_handle_valid()
Create function _client_handle_valid() to encapsulate checking that
a client handle value is in range, and has a valid endpoint.  Have
error messages distinguish between handles that are out-of-range
and those whose endpoints are marked invalid.

Also define macro client_handle_valid() whose purpose it is to
supply the name of the calling function to the function.  Use the
macro in ipa3_cfg_ep_seq().

Change-Id: I1afd278280323219b0d229c6137cd7f70d38ee66
Signed-off-by: Alex Elder <elder@codeaurora.org>