msm: ipa: remove some client_handle_valid() calls
Quite a few functions in "ipa_utils.c" are now private, and are only
called once, by ipa3_cfg_ep().  That function starts by validating
the client handle, so we know the client handle is valid for all
functions it calls.  We can therefore drop the handle validation
checks in the called functions.

Change-Id: I925f7bd54ab6a7a40ec9ce42f6862f6323a9c5fb
Signed-off-by: Alex Elder <elder@codeaurora.org>