msm: ipa: drop a parameter check of ipa3_set_rt_tuple_mask()
The one caller of ipa3_set_rt_tuple_mask() is in the IPA code
itself.  The call is in the body of a for loop, which iterates with
the "tbl_idx" value ranging from 0 to (less than) this limit:
   max(ipa3_mem(V6_RT_NUM_INDEX), ipa3_mem(V4_RT_NUM_INDEX));

The first check of "tbl_idx" in ipa3_set_rt_tuple_mask() returns a
failure if:
  - it is max(ipa3_mem(V6_RT_NUM_INDEX), ipa3_mem(V4_RT_NUM_INDEX))
    or more (which it never will be)
  - it is less than 0 (again, which it never will be)

There is no need for ipa3_set_rt_tuple_mask() to re-verify these
condtions, so that check can be removed.

Change-Id: Idbd3dd03f4052e123c9d3d536f2e585ee0f1209f
Signed-off-by: Alex Elder <elder@codeaurora.org>