msm: ipahal: have ipa_write_64() return no value
The comments above ipa_write_64() claim it returns the converted
value (or its address), but in fact what's returned is an address
*beyond* the converted value in memory.

That function is only used as the "write_val_to_hdr" function for
a IPA HAL filter/route object, and none of the callers of that
function ever check its return value.  So just have that function
return no value, and update the definition of the write_val_to_hdr
field accordingly.

Change-Id: Ieb6bef7e98646016cab484a8a5813c0d7ab32775
Signed-off-by: Alex Elder <elder@codeaurora.org>