msm: ipa: dest is non-null in ipa_write_64()
The only use of ipa_write_64() is as a "write_val_to_hdr" function
for an IPA HAL filter/route object.  ipahal_rt_generate_empty_img()
and ipahal_flt_generate_empty_img() are the only functions that call
that function, and they are guaranteed to be supplying a non-null
"dest" pointer (it was allocated in calling function).  So we can
get rid of the check for non-null "dest" in ipa_write_64().

Change-Id: I43072484e41181e1ea34b894f83d0a2a552326b7
Signed-off-by: Alex Elder <elder@codeaurora.org>