msm: ipa: introduce ipa_client_string()
Create a new function ipa_client_string() to return the name of a
given client as a constant string.  This hides the details of how
the name is derived and allows us to make the client strings table
private.  Use this function in the expansion of the macro
IPA_ACTIVE_CLIENTS_PREP_EP().

Change-Id: Ifea3a4829e7c56bf246ed82ffe9057ddaa778a4e
Signed-off-by: Alex Elder <elder@codeaurora.org>