msm: rmnet_ipa: kill dev_num
The "dev_num" global is never initialized, but is passed as the
count value to unregister_chrdev_region() in ipa3_wan_ioctl_deinit().
This is an error, because the region was never registered.  (It
turns out this is currently harmless because both values passed to
unregister_chrdev_region() are 0.)  Remove "dev_num" and the
function call that uses it.

Change-Id: Ieec81ff965068972bb882134bf6cc5331ced0066
Signed-off-by: Alex Elder <elder@codeaurora.org>