msm: rmnet_ipa: clean up ipa3_wwan_probe() error paths
If an error is returned when register_netdev() is called by
ipa3_wwan_probe(), the function returns an error without properly
cleaning up previously-set state.  The activities that occur at the
config_err label are the things we need, so jump there in case of
error.  There is a block of code just below that that will never be
taken because we can see that the value of err will be 0; so delete
that conditional block.

Change-Id: I6e00def16e3caff7dfb5518289382ca2be2d9647
Signed-off-by: Alex Elder <elder@codeaurora.org>