msm: rmnet_ipa: new_mtu is non-negative in ipa3_wwan_change_mtu()
The value of new_mtu in ipa3_wwan_change_mtu() will never be
negative.  It is called only as the net_device_ops->ndo_change_mtu
method, which is only called by __dev_set_mtu(), which is in turn
only called by dev_set_mtu().  And dev_set_mtu() returns early with
an error if the MTU value provided was negative.

Do a little reformatting in the function as well.

Change-Id: I9aef8e456505f48a19899bd8e65c24182ecf266b
Signed-off-by: Alex Elder <elder@codeaurora.org>