msm: ipa_qmi: handle is good in ipa3_a5_svc_disconnect_cb()
There is no need to verify that the handle provided to
ipa3_a5_svc_disconnect_cb() is the same as ipa3_svc_handle.

That function is only called as a disconnect_cb function for
ipa3_a5_svc_ops_options.  The disconnect_cb function is only called
by handle_rmv_client().  And the only reason that function would
call this particular disconnect function is if it was associated
with the handle that's supplied to it.  So the handle supplied
to ipa3_a5_svc_disconnect_cb() *will* be ipa3_svc_handle.

Do a little minor formatting of the function as well.

Change-Id: Iaf329a6151c88a19cd77cc7b78fd4d568533218a
Signed-off-by: Alex Elder <elder@codeaurora.org>